+++ clients/privacy-center/pages/api/fides-js.ts
@@ -347,3 +347,9 @@ async function fetchCustomFidesCss(
   }
   return cachedCustomFidesCss;
 }
+
+export const config = {
+  api: {
+    responseLimit: false,
+  },
+};